Allow classes to register builder functions
Some classes should have required constructor arguments. Allow
these classes to define a builder other than the constructor
which will pass default arguments to the constructor. This allows
the constructor to force arguments when used by a programmer, while
allowing the Document to construct the object as RDF is read
from file.