Clean up singleton properties on read
When reading a file, we can end up with multiple values in singleton
properties due to defaulting values. Clean these up in a post-read
phase.

Yes, this is a bit messy. We have a choice between validating objects
in their constructor, which is good for a user, and delaying validation
which is good for the document reader. We're leaning towards the former
here, and allowing the document read to be a bit messy.