Change constructor arg "name" to "identity"
The use of "name" as the argument caused confusion. The argument does not
populate the "name" field of an object. Rather, it is used to populate the
identity and/or display_id. Switch to calling it identity, which is the
first use in the heuristic. If the value is not a valid identity it is
assumed to be a display_id and concatenated with a default namespace if
one exists.