Fix compliant name cascading
If an owned object is added and given a compliant name, and it has
child owned objects that need a compliant name, and one of those
child objects has a type_uri outside the SBOL3 namespace (like Measure)
then the child object may be given an invalid displayId. Fix that
by properly parsing the child object's type_uri.