Fix document reading with default namespace
Gracefully handle document reading when the default namespace is set
instead of erroring out. Do a better job of choosing an object's
namespace when the identity is a fully specified URL. This case allows
ignoring the default namespace if it does not match the identity URL.

Add several new unit tests for various cases, and update other unit
tests based on this change.