Incorporate RDFLib JSON-LD fix
RDFLib has fixed the JSON-LD parser bug that adds a trailing slash to
a URL in that is an "object" in a s-p-o triple. Bump the RDFLib
dependency to the version that includes the fix, and remove the
expected failure annotation on the unit test.

See https://github.com/RDFLib/rdflib/issues/1443