Add `Document.migrate`
`Document.migrate` _moves_ objects into the invoked document. It does
not matter if the objects are currently members of a document or
not. For objects that are not currently in a document this is
equivalent to `Document.add()`. For documents that are in a document,
they are removed from that document and added to this one.