Accelerate loading
Searching for orphans when loading large documents is incredibly slow,
taking 90+% of the load time. Optimize the search for orphans by
tracking the child objects that get assigned to parent objects. Use
this info to check if a candidate orphan is in the set of assigned
child objects. This reduces load time by 98% in the test document
provided by iGEM.