Support copy of custom classes
Copying a CustomTopLevel or CustomIdentified without a registered builder would cause an error.
Instead, we can fall back onto using a CustomTopLevel or CustomIdentified, just as we do when loading the document in the first place.
Test has been extended to check the case of unregistered custom objects for both round tripping and copying